Add ‘Paste’ command to Sketch plugin
