+++ asketch2sketch/asketch2sketch.js
@@ -94,43 +94,14 @@ function addSharedColor(document, colorJSON) {
   assets.addColor(color);
 }
 
-export default function asketch2sketch(context) {
+export default function asketch2sketch(context, asketchFiles) {
   const document = context.document;
   const page = document.currentPage();
 
   let asketchDocument = null;
   let asketchPage = null;
 
-  const panel = NSOpenPanel.openPanel();
-
-  panel.setCanChooseDirectories(false);
-  panel.setCanChooseFiles(true);
-  panel.setAllowsMultipleSelection(true);
-  panel.setTitle('Choose *.asketch.json files');
-  panel.setPrompt('Choose');
-  panel.setAllowedFileTypes(['json']);
-
-  if (panel.runModal() !== NSModalResponseOK || panel.URLs().length === 0) {
-    return;
-  }
-
-  const urls = panel.URLs();
-
-  urls.forEach(url => {
-    const data = NSData.dataWithContentsOfURL(url);
-    const content = NSString.alloc().initWithData_encoding_(data, NSUTF8StringEncoding);
-
-    let asketchFile = null;
-
-    try {
-      asketchFile = JSON.parse(content);
-    } catch (e) {
-      const alert = NSAlert.alloc().init();
-
-      alert.setMessageText('File is not a valid JSON.');
-      alert.runModal();
-    }
-
+  asketchFiles.forEach(asketchFile => {
     if (asketchFile && asketchFile._class === 'document') {
       asketchDocument = asketchFile;
     } else if (asketchFile && asketchFile._class === 'page') {
