+++ asketch2sketch/manifest.json
@@ -4,16 +4,25 @@
   "icon": "icon.png",
   "commands": [
     {
-      "name": "From *Almost* Sketch to Sketch",
-      "identifier": "asketch2sketch",
-      "script": "./asketch2sketch.js"
+      "name": "Import file(s)...",
+      "identifier": "importFiles",
+      "script": "./importFiles.js"
+    },
+    {
+      "name": "Paste",
+      "identifier": "paste",
+      "script": "./paste.js"
     }
   ],
   "menu": {
     "isRoot": true,
     "title": "output",
-    "items": [
-      "asketch2sketch"
-    ]
+    "items": [{
+      "title": "From *Almost* Sketch to Sketch",
+      "items": [
+        "importFiles",
+        "paste"
+      ]
+    }]
   }
 }
